==> notebooks/3-feature-engineering-and-preprocessing.ipynb <==
%matplotlib inline
import numpy as np
import pandas as pd
import pandas_profiling
import matplotlib.pyplot as plt
import seaborn as sns
from collections import OrderedDict
import xgboost as xgb
import lightgbm as lgb
sns.set_context('poster')

from sklearn.preprocessing import MinMaxScaler
from sklearn.ensemble import RandomForestRegressor, GradientBoostingRegressor
from sklearn.linear_model import Ridge, Lasso, LinearRegression
from sklearn.model_selection import train_test_split

# Importing from my own modules
import sys
sys.path.append('../financial_forecasting/')
from utils import load_data, wMSE, train_and_test_models
from preprocessing import Imputer, LogTransformer, MeanEncoder, compute_combined_variable, TreeBinner

# Load data

df_train, df_test = load_data()

X = df_train.drop(labels=['y','Weight'], axis=1)
y = df_train.y
weights = df_train.Weight
X_test = df_test

np.random.seed=42
X_train, X_val, y_train, y_val = train_test_split(X,y,test_size=0.30)

weights_train = weights[X_train.index]
weights_val = weights[X_val.index]

print(X_train.shape)
print(y_train.shape)

print(X_val.shape)
print(y_val.shape)

print(X_test.shape)

# Apply pre-processing

# Missing value imputation
imputer = Imputer()
imputer.fit(X_train)
X_train = imputer.transform(X_train)
X_val = imputer.transform(X_val)
X_test = imputer.transform(X_test)

# Log transformation
features_to_log_trans = ['x0', 'x1', 'x2', 'x3A', 'x3B', 'x3C', 'x3D', 'x3E', 'x4', 'x5', 'x6']
log_transformer = LogTransformer()
log_transformer.fit(X_train, features=features_to_log_trans)
X_train = log_transformer.transform(X_train)
X_val = log_transformer.transform(X_val)
X_test = log_transformer.transform(X_test)

print(len(X))
print(len(X_train)+len(X_val)+len(X_test))

# Compute log differences between current day and previous day for each stock
# Need to combine train, validation, and test for this
X_train['is_train_val_test'] = 0
X_val['is_train_val_test'] = 1
X_test['is_train_val_test'] = 2

X = pd.concat([X_train, X_val, X_test])
X = X.reset_index(drop=True)
X = compute_combined_variable(X, var1='Day', var2='Stock')
features = [feat + '_log10' for feat in df_train.columns if 'x' in feat]
for feat in features:
    x = X.groupby(['Day','Stock'])[feat].mean().unstack().diff().fillna(0).stack().reset_index()
    x = compute_combined_variable(x, var1='Day', var2='Stock')
    log_diff_dict = x.set_index('Day_Stock')[0].to_dict()
    X[feat + '_diff'] = X.loc[:, 'Day_Stock'].map(log_diff_dict)

# Split back into train, validation, and test
X_train = X[X.is_train_val_test == 0]
X_val = X[X.is_train_val_test == 1]
X_test = X[X.is_train_val_test == 2]

# Tree based binning
features_to_bin = features_to_log_trans.copy()
tree_binner = TreeBinner()
tree_binner.fit(X_train, y_train, weights_train, features=features_to_bin)
X_train = tree_binner.transform(X_train)
X_val = tree_binner.transform(X_val)
X_test = tree_binner.transform(X_test)

# Mean value encoding
features_to_encode = ['Market', 'Day', 'Stock']
encoder = MeanEncoder()
encoder.fit(X_train, features=features_to_encode, target=y_train)
X_train = encoder.transform(X_train)
X_val = encoder.transform(X_val)
X_test = encoder.transform(X_test)

# Drop columns no longer needed
X_train.drop(labels=['is_train_val_test', 'Day_Stock'], axis=1, inplace=True)
X_val.drop(labels=['is_train_val_test', 'Day_Stock'], axis=1, inplace=True)
X_test.drop(labels=['is_train_val_test', 'Day_Stock'], axis=1, inplace=True)

# Output processed data

X_train.to_csv('../data/preprocessed/train.csv', index=False)
X_val.to_csv('../data/preprocessed/validation.csv', index=False)
X_test.to_csv('../data/preprocessed/test.csv', index=False)

weights_train.to_csv('../data/preprocessed/train_weights.csv', index=False)
weights_val.to_csv('../data/preprocessed/validation_weights.csv', index=False)

y_train.to_csv('../data/preprocessed/train_target.csv', index=False)
y_val.to_csv('../data/preprocessed/validation_target.csv', index=False)


# All the features
enabled_vars_trees = X_train.columns

models = OrderedDict([
                      ('lgboost1', lgb.LGBMRegressor(n_estimators=100, n_jobs=-1, learning_rate=0.1)), 
                    ])

# Remove zero feature importance features
df_preds_train, df_preds_test, train_error, test_error = train_and_test_models(models, 
                                                         X_train.loc[:,enabled_vars_trees], y_train, 
                                                         X_val.loc[:,enabled_vars_trees], y_val, 
                                                         weights_train, weights_val)

feature_importance = pd.Series(models['lgboost1'].feature_importances_, index=enabled_vars_trees)
feature_importance = np.abs(feature_importance) / np.abs(feature_importance).sum()
print(feature_importance.sort_values(ascending=False))
non_zero_feature_importance = set(feature_importance[feature_importance > 1e-5].index)

# All the features
enabled_vars_trees = non_zero_feature_importance

models = OrderedDict([
                      ('lgboost1', lgb.LGBMRegressor(n_estimators=100, n_jobs=-1, learning_rate=0.1)), 
                    ])

# Remove zero feature importance features
df_preds_train, df_preds_test, train_error, test_error = train_and_test_models(models, 
                                                         X_train.loc[:,enabled_vars_trees], y_train, 
                                                         X_val.loc[:,enabled_vars_trees], y_val, 
                                                         weights_train, weights_val)

best_feats_so_far = ['Day',
 'Day_mean_encoded',
 'Market',
 'Market_mean_encoded',
 'Stock',
 'Stock_mean_encoded',
 'x0',
 'x0_log10',
 'x1',
 'x1_binned',
 'x2_log10',
 'x3A',
 'x3A_log10',
 'x3B',
 'x3C',
 'x3C_log10',
 'x3D',
 'x3D_log10',
 'x3E',
 'x3E_binned',
 'x3E_log10',
 'x4',
 'x4_binned',
 'x4_log10',
 'x5',
 'x5_log10',
 'x6',
 'x6_binned']

set(best_feats_so_far).issubset(set(non_zero_feature_importance))

# Best features found by hand
enabled_vars_trees = best_feats_so_far

models = OrderedDict([
                      ('lgboost1', lgb.LGBMRegressor(n_estimators=100, n_jobs=-1, learning_rate=0.1)), 
                    ])

# Remove zero feature importance features
df_preds_train, df_preds_test, train_error, test_error = train_and_test_models(models, 
                                                         X_train.loc[:,enabled_vars_trees], y_train, 
                                                         X_val.loc[:,enabled_vars_trees], y_val, 
                                                         weights_train, weights_val)

# Modelling

## Training and testing

### Null model

class NullModel():
    def __init__(self):
        self.stock_mean = None
    
    def fit(self, df):
        stock_mean = df.groupby('Stock')['y'].mean()
        self.stock_mean = stock_mean
        
    def predict(self, df):
        preds = df['Stock'].map(self.stock_mean).rename('y')
        preds.fillna(0)
        return preds

N = 436671
X_train_ = X_train.iloc[:N, :]
y_train_ =  y_train.iloc[:N]
weights_train_ = weights_train.iloc[:N]
df = pd.concat([X_train_, y_train_], axis=1)

clfs = [NullModel()]

for clf in clfs:
    # Training
    clf.fit(df)
        
    # Testing
    preds_train = clf.predict(X_train_)
    preds_val = clf.predict(X_val)
    preds_test = clf.predict(X_test)
    
    train_error = wMSE(preds=preds_train, y=y_train_, weights=weights_train_)
    val_error =  wMSE(preds=preds_val, y=y_val, weights=weights_val)
    scale = len(df_test)
    print('Train error: {} Test error: {} \n'.format(train_error * scale, val_error * scale))

### Linear models

N = 436671
X_train_ = X_train_linear.iloc[:N, :]
y_train_ =  y_train.iloc[:N]
weights_train_ = weights_train.iloc[:N]

clfs = {'LinearRegression':LinearRegression(), 
        'L1Regression':Lasso(),
        'L2Regression':Ridge(alpha=1)}

preds_v = pd.DataFrame()
preds_t = pd.DataFrame()

for clf_name, clf in clfs.items():
    # Training
    print('Fitting: {}'.format(clf))
    try:
        clf.fit(X_train_, y_train_, sample_weight=weights_train_.values)
    except TypeError:
        print('{} does not accept sample weights'.format(clf))
        clf.fit(X_train_, y_train_)
        
    # Testing
    preds_train = clf.predict(X_train_)
    preds_val = clf.predict(X_val_linear)
    preds_test = clf.predict(X_test_linear)
    
    train_error = wMSE(preds=preds_train, y=y_train_, weights=weights_train_)
    val_error =  wMSE(preds=preds_val, y=y_val, weights=weights_val)
    scale = len(df_test)
    print('Train error: {} Test error: {} \n'.format(train_error * scale, val_error * scale))
    
    # Append test predictions to a dataframe
    data = {clf_name + '_preds_val': preds_val}
    df_preds_val = pd.DataFrame(data=data, index=X_val_linear.index)
    preds_v = pd.concat([preds_v, df_preds_val], axis=1)
    
    # Append test predictions to a dataframe
    data = {clf_name + '_preds_test': preds_test}
    df_preds_test = pd.DataFrame(data=data, index=X_test_linear.index)
    preds_t = pd.concat([preds_t, df_preds_test], axis=1)

### Tree based models

N = 436671
X_train_ = X_train.iloc[:N, :]
y_train_ =  y_train.iloc[:N]
weights_train_ = weights_train.iloc[:N]

clfs = {'RF': RandomForestRegressor(n_estimators=100),
        'GBM': GradientBoostingRegressor()}
for clf_name, clf in clfs.items():
    # Training
    print('Fitting: {}'.format(clf))
    try:
        clf.fit(X_train_, y_train_, sample_weight=weights_train_.values)
    except TypeError:
        print('{} does not accept sample weights'.format(clf))
        clf.fit(X_train_, y_train_)
        
    # Testing
    preds_train = clf.predict(X_train_)
    preds_val = clf.predict(X_val)
    preds_test = clf.predict(X_test)
    
    train_error = wMSE(preds=preds_train, y=y_train_, weights=weights_train_)
    val_error =  wMSE(preds=preds_val, y=y_val, weights=weights_val)
    scale = len(df_test)
    print('Train error: {} Test error: {} \n'.format(train_error * scale, val_error * scale))
    
    # Append test predictions to a dataframe
    data = {clf_name + '_preds_val': preds_val}
    df_preds_val = pd.DataFrame(data=data, index=X_val.index)
    preds_v = pd.concat([preds_v, df_preds_val], axis=1)
    
    # Append test predictions to a dataframe
    data = {clf_name + '_preds_test': preds_test}
    df_preds_test = pd.DataFrame(data=data, index=X_test.index)
    preds_t = pd.concat([preds_t, df_preds_test], axis=1)

### Model Stacking

preds_v['simple_average'] = (preds_v.L2Regression_preds_val + preds_v.RF_preds_val + preds_v.LinearRegression_preds_val)/3
preds_t['simple_average'] = (preds_t.L2Regression_preds_test + preds_t.RF_preds_test + preds_t.LinearRegression_preds_test)/3

error = wMSE(preds=preds_v.simple_average, y=y_val, weights=weights_val)
print(error*scale)

preds_t['simple_average'].rename('y').to_csv('../data/output.csv', header=True)

# Learning curve
train_errors = []
test_errors = []
data_size = []
for N in range(1000,42000,5000):
    X_train_ = X_train.iloc[:N, :]
    y_train_ =  y_train.iloc[:N]
    weights_train_ = weights_train.iloc[:N]

    clf = RandomForestRegressor(n_estimators=100)
    clf.fit(X_train_, y_train_, sample_weight=weights_train_.values)
    
    preds_train = clf.predict(X_train_)
    preds_test = clf.predict(X_test)
    
    train_error = wMSE(preds=preds_train, y=y_train_, weights=weights_train_)
    test_error =  wMSE(preds=preds_test, y=y_test, weights=weights_test)
    
    print('Data size: {} Train error: {} Test error: {}'.format(N,train_error * 1e6, test_error*1e6))
    
    train_errors.append(train_error)
    test_errors.append(test_error)
    data_size.append(N)
    
plt.plot(data_size, train_errors)
plt.plot(data_size, test_errors)

# Sanity check
df_train_error = pd.DataFrame({'preds_train':preds_train, 'y_train':y_train, 'weights_train':weights_train})
df_train_error['Diff'] = df_train_error.preds_train - df_train_error.y_train
df_train_error['wMSE'] = np.square(df_train_error.Diff) * df_train_error.weights_train
print(df_train_error.wMSE.sum() / len(df_train_error))

## Feature importance

feat_importance = pd.Series(dict(zip(X_train.columns, clf.feature_importances_))).sort_values(ascending=False)
feat_importance[feat_importance > 1e-4].plot(kind='barh')

clf.get_params()

# Output 

preds_t['simple_average'].rename('y').to_csv('../data/output.csv', header=True)
