==> notebooks/9.0-use-entity-embeddings-with-gbt.ipynb <==
%matplotlib inline
import numpy as np
import pandas as pd
import matplotlib.pyplot as plt
import seaborn as sns
from collections import OrderedDict
sns.set_context('poster')

from sklearn.preprocessing import MinMaxScaler, StandardScaler
from sklearn.ensemble import RandomForestRegressor, GradientBoostingRegressor
from sklearn.linear_model import Ridge, Lasso, LinearRegression
from sklearn.model_selection import train_test_split
from sklearn.cross_validation import cross_val_score
from sklearn.metrics import make_scorer
from sklearn.externals import joblib
import xgboost as xgb
import lightgbm as lgb

# Importing from my own modules
import sys
sys.path.append('../financial_forecasting/')
from utils import load_data, wMSE, train_and_test_models
from preprocessing import Imputer, LogTransformer, MeanEncoder, TreeBinner, transform_to_embedding_vec

# Load data

X_train = pd.read_csv('../data/preprocessed/train.csv')
X_val = pd.read_csv('../data/preprocessed/validation.csv')
X_test = pd.read_csv('../data/preprocessed/test.csv')

weights_train = pd.read_csv('../data/preprocessed/train_weights.csv', squeeze=True)
weights_val = pd.read_csv('../data/preprocessed/validation_weights.csv', squeeze=True)

y_train = pd.read_csv('../data/preprocessed/train_target.csv', squeeze=True)
y_val = pd.read_csv('../data/preprocessed/validation_target.csv', squeeze=True)

# Load embeddings

emb_file = '../data/embeddings.pkl'
embedding_matrix = joblib.load(emb_file)

df_embeddings_train = transform_to_embedding_vec(X_train.Stock, embedding_matrix)
df_embeddings_val = transform_to_embedding_vec(X_val.Stock, embedding_matrix)
df_embeddings_test = transform_to_embedding_vec(X_test.Stock, embedding_matrix)

X_train = pd.concat([X_train, df_embeddings_train], axis=1)
X_val = pd.concat([X_val, df_embeddings_val], axis=1)
X_test = pd.concat([X_test, df_embeddings_test], axis=1)

# Train and evaluate models using embeddings

embedding_feats = list(df_embeddings_train.columns)

feats_no_embeddings = ['Day', 'Market', 'Market_mean_encoded', 'Stock', 'Stock_mean_encoded',
                       'x0', 'x0_log10', 'x0_log10_diff', 'x1_log10', 'x1_log10_diff',
                       'x2_log10', 'x2_log10_diff', 'x3A', 'x3A_log10', 'x3A_log10_diff',
                       'x3B', 'x3B_binned', 'x3B_log10', 'x3C', 'x3C_log10', 'x3D',
                       'x3D_log10', 'x3D_log10_diff', 'x3E', 'x3E_log10', 'x3E_log10_diff',
                       'x4', 'x4_binned', 'x4_log10_diff', 'x5', 'x5_binned', 'x5_log10',
                       'x5_log10_diff', 'x6', 'x6_binned', 'x6_log10_diff']

feats_with_embeddings = feats_no_embeddings + embedding_feats

enabled_vars = feats_no_embeddings

models = OrderedDict([
                      ('lgboost_reg', lgb.sklearn.LGBMRegressor(n_estimators=350, n_jobs=-1, max_depth=5, reg_lambda=10.0)),
                      ('lgboost', lgb.sklearn.LGBMRegressor(n_estimators=700)),
                    ])

df_preds_train, df_preds_test, train_error, test_error = train_and_test_models(models, 
                                                         X_train.loc[:,enabled_vars], y_train, 
                                                         X_val.loc[:,enabled_vars], y_val, 
                                                         weights_train, weights_val)

enabled_vars = feats_with_embeddings

models = OrderedDict([
                      ('lgboost_reg', lgb.sklearn.LGBMRegressor(n_estimators=350, n_jobs=-1, max_depth=5, reg_lambda=10.0)),
                      ('lgboost', lgb.sklearn.LGBMRegressor(n_estimators=700)),
                    ])

df_preds_train, df_preds_test, train_error, test_error = train_and_test_models(models, 
                                                         X_train.loc[:,enabled_vars], y_train, 
                                                         X_val.loc[:,enabled_vars], y_val, 
                                                         weights_train, weights_val)

Embeddings have definitely helped compared to using no embeddings. 

# Train final model

enabled_vars = feats_with_embeddings

models = OrderedDict([
                      ('xgboost', xgb.sklearn.XGBRegressor(n_estimators=700, n_jobs=-1)), 
                      ('lgboost_reg', lgb.sklearn.LGBMRegressor(n_estimators=350, n_jobs=-1, max_depth=5, reg_lambda=10.0)),
                      ('lgboost', lgb.sklearn.LGBMRegressor(n_estimators=700)),
                      ('xgboost_reg', xgb.sklearn.XGBRegressor(n_estimators=700, n_jobs=-1, reg_lambda=10.0, grow_policy='lossguide', tree_method='hist', max_depth=5)), 
                    ])

df_preds_train, df_preds_test, train_error, test_error = train_and_test_models(models, 
                                                         X_train.loc[:,enabled_vars], y_train, 
                                                         X_val.loc[:,enabled_vars], y_val, 
                                                         weights_train, weights_val)

# Model stacking
# Validation set
preds = (df_preds_test.xgboost_preds_test + 
         df_preds_test.lgboost_preds_test +
         df_preds_test.xgboost_reg_preds_test +
         df_preds_test.lgboost_reg_preds_test
        )/4
wMSE(preds=preds, y=y_val, weights=weights_val)

Score on validation set with entity embeddings is better.

# Model stacking
# Test set
p1 = models['xgboost'].predict(X_test.loc[:, enabled_vars])
p2 = models['lgboost'].predict(X_test.loc[:, enabled_vars])
p3 = models['lgboost_reg'].predict(X_test.loc[:, enabled_vars])
p4 = models['xgboost_reg'].predict(X_test.loc[:, enabled_vars])

preds_test = pd.DataFrame(np.c_[p1,p2,p3,p4], columns=['xgb','lgb', 'lgbr', 'xgbr'], index=X_test.index)

preds_test['y'] = preds_test.mean(axis=1)

preds_test.head()

preds_test.y.to_csv('../data/output_13_lgb350reg_lgb700_xgb700_xgb700reg_with_log_diff_as_feats_with_param_tuning_entity_embeddings.csv', index=True, header=True)

# Train final model with all the data

X_train = pd.concat([X_train, X_val])
y_train = pd.concat([y_train, y_val])
weights_train = pd.concat([weights_train, weights_val])

X_train.reset_index(drop=True, inplace=True)
y_train.reset_index(drop=True, inplace=True)
weights_train.reset_index(drop=True, inplace=True)

enabled_vars = feats_with_embeddings

models = OrderedDict([
                      ('lgboost_reg', lgb.sklearn.LGBMRegressor(n_estimators=350, n_jobs=-1, max_depth=5, reg_lambda=10.0)),
                      ('lgboost', lgb.sklearn.LGBMRegressor(n_estimators=700)),
                      ('xgboost_reg', xgb.sklearn.XGBRegressor(n_estimators=700, n_jobs=-1, reg_lambda=10.0, grow_policy='lossguide', tree_method='hist', max_depth=5)), 
                      ('xgboost', xgb.sklearn.XGBRegressor(n_estimators=700, n_jobs=-1)),
                    ])

df_preds_train, df_preds_test, train_error, test_error = train_and_test_models(models, 
                                                         X_train.loc[:,enabled_vars], y_train, 
                                                         X_val.loc[:,enabled_vars], y_val, 
                                                         weights_train, weights_val)

enabled_vars = feats_with_embeddings

models = OrderedDict([
                      ('xgboost', xgb.sklearn.XGBRegressor(n_estimators=700, n_jobs=-1)), 
                      ('lgboost_reg', lgb.sklearn.LGBMRegressor(n_estimators=350, n_jobs=-1, max_depth=5, reg_lambda=10.0)),
                      ('lgboost', lgb.sklearn.LGBMRegressor(n_estimators=700)),
                      ('xgboost_reg', xgb.sklearn.XGBRegressor(n_estimators=700, n_jobs=-1, reg_lambda=10.0, grow_policy='lossguide', tree_method='hist', max_depth=5)), 
                    ])

df_preds_train, df_preds_test, train_error, test_error = train_and_test_models(models, 
                                                         X_train.loc[:,enabled_vars], y_train, 
                                                         X_val.loc[:,enabled_vars], y_val, 
                                                         weights_train, weights_val)

# Model stacking
# Validation set
preds = (df_preds_test.xgboost_preds_test + 
         df_preds_test.lgboost_preds_test +
         df_preds_test.xgboost_reg_preds_test +
         df_preds_test.lgboost_reg_preds_test
        )/4
wMSE(preds=preds, y=y_val, weights=weights_val)

# Model stacking
# Test set
p1 = models['xgboost'].predict(X_test.loc[:, enabled_vars])
p2 = models['lgboost'].predict(X_test.loc[:, enabled_vars])
p3 = models['lgboost_reg'].predict(X_test.loc[:, enabled_vars])
p4 = models['xgboost_reg'].predict(X_test.loc[:, enabled_vars])

preds_test = pd.DataFrame(np.c_[p1,p2,p3,p4], columns=['xgb','lgb', 'lgbr', 'xgbr'], index=X_test.index)

preds_test['y'] = preds_test.mean(axis=1)

preds_test.head()

preds_test.y.to_csv('../data/output_14_lgb350reg_lgb700_xgb700_xgb700reg_with_log_diff_as_feats_with_param_tuning_entity_embeddings_all_data_2.csv', index=True, header=True)
